+++ lib/monolens/core/mapping.rb
@@ -4,7 +4,7 @@ module Monolens
       include Lens
 
       def call(arg, world = {})
-        option(:values, {}).fetch(arg) do
+        option(:defn, option(:values, {})).fetch(arg) do
           on_missing(arg, world)
         end
       end
@@ -14,19 +14,19 @@ module Monolens
       def on_missing(arg, world)
         strategy = option(:on_missing, :fail)
         case strategy.to_sym
-        when :keep
-          arg
-        when :fail
-          fail!("Unrecognized value `#{arg}`", world)
         when :default
           option(:default, nil)
-        when :null
-          nil
+        when :fail
+          fail!("Unrecognized value `#{arg}`", world)
         when :fallback
           missing_fallback = ->(arg, world) do
             raise Monolens::Error, "Unexpected missing fallback handler"
           end
           option(:fallback, missing_fallback).call(self, arg, world)
+        when :keep
+          arg
+        when :null
+          nil
         else
           raise Monolens::Error, "Unexpected missing strategy `#{strategy}`"
         end