+++ spec/monolens/core/test_mapping.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe Monolens, 'core.mapping' do
   let(:mapping) do
-    { 'values' => { 'todo' => 'open' } }
+    { 'defn' => { 'todo' => 'open' } }
   end
 
   context 'with default options' do
@@ -102,6 +102,28 @@ describe Monolens, 'core.mapping' do
 
     it 'sets the location correctly' do
       expect(subject.location).to eql([1])
+    end
+  end
+
+  context 'backward compatibility' do
+    let(:mapping) do
+      { 'values' => { 'todo' => 'open' } }
+    end
+
+    context 'with using values instead of defn' do
+      subject do
+        Monolens.lens('core.mapping' => mapping)
+      end
+
+      it 'replaces the value by its mapped' do
+        expect(subject.call('todo')).to eql('open')
+      end
+
+      it 'raises if not found' do
+        expect {
+          subject.call('nosuchone')
+        }.to raise_error(Monolens::LensError)
+      end
     end
   end
 end