+++ spec/test_documentation.rb
@@ -29,18 +29,24 @@ def yaml_load(input)
 end
 
 Path.dir.parent.glob('documentation/*/*.md') do |file|
-  describe file do
-    if file.read =~ rx
-      lens, input, output = $1, $2, $3
+  #next unless file.to_s =~ /literal/
 
-      it 'has correct examples' do
-        lens = Monolens.lens(yaml_load(lens))
-        input = yaml_load(input)
-        output = yaml_load(output)
-        expect(lens.call(input)).to eql(output)
-      end
-    else
+  describe file do
+    matches = file.read.scan(rx)
+
+    if matches.empty?
       puts "WARN: missing or invalid example in #{file}"
+    else
+      matches.each do |example|
+        lens, input, output = example
+
+        it 'has correct examples' do
+          lens = Monolens.lens(yaml_load(lens))
+          input = yaml_load(input)
+          output = yaml_load(output)
+          expect(lens.call(input)).to eql(output)
+        end
+      end
     end
   end
 end