+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   used to write singular object lenses and easily test them
   against .csv and .xlsx files.
 
+* core.mapping now supports a :keep on_error strategy that
+  simply return the input key.
+
 * Add missing binary in the ruby gem.
 
 ## 0.5.3 - 2022-06-09